test(radio): migrate color
