+++ core/src/components/radio/test/color/radio.e2e.ts-snapshots/radio-color-checked-ios-ltr-Mobile-Safari-linux.png
