+++ core/src/components/radio/test/color/radio.e2e.ts-snapshots/radio-color-unchecked-ios-ltr-Mobile-Chrome-linux.png
