+++ core/src/components/radio/test/color/radio.e2e.ts-snapshots/radio-color-unchecked-md-ltr-Mobile-Firefox-linux.png
